requirements.txt
lxml
pillow
requests
pyYAML
future
